[refactor] Project 상세 페이지 Header 분리
- Project 상세 페이지 Header 분리
  기존 header는 title만 보여줘서 너무 작은 component였음
  Project Detail에 있던 header의 Edit 상태와 상태 변경 함수를 모두 Header Component로 이동
